уточнение типов данных

+++ freeLib/src/importthread.cpp
@@ -30,7 +30,7 @@ void ClearLib(QSqlDatabase dbase, qlonglong id_lib, bool delete_only)
     }
 }
 
-void GetBookInfo(book_info &bi,const QByteArray &data,QString type,bool info_only,qlonglong id_book)
+void GetBookInfo(book_info &bi,const QByteArray &data,QString type,bool info_only,int id_book)
 {
     bi.id=id_book;
     if(id_book<0 || !info_only)
